+++ Changelog.md
@@ -1,6 +1,6 @@
 # Goldfish Changelog
 
-## 0.2.2 (unreleased)
+## 0.2.2 (2015-06-04)
 
 Features: